requirements.txt
selenium
waveshare-epaper